feat(images): load images via zustand store
