+++ src/pages/Images.tsx
@@ -1,26 +1,9 @@
-import { useEffect, useState } from "react";
 import PaginatedImages from "../components/PaginatedImages";
-
-interface ImageData {
-  pixels: number[];
-  label: number;
-}
+import { useImageStore } from "../stores/useImageStore";
 
 export default function Images() {
-  const [trainImages, setTrainImages] = useState<ImageData[]>([]);
-  const [valImages, setValImages] = useState<ImageData[]>([]);
-
-  useEffect(() => {
-    const load = async () => {
-      const res = await fetch("/digits_8x8.json");
-      const data: ImageData[] = await res.json();
-      const shuffled = data.sort(() => Math.random() - 0.5);
-      const split = Math.floor(shuffled.length * 0.8);
-      setTrainImages(shuffled.slice(0, split));
-      setValImages(shuffled.slice(split));
-    };
-    void load();
-  }, []);
+  const trainImages = useImageStore((s) => s.trainImages);
+  const valImages = useImageStore((s) => s.valImages);
 
   return (
     <div className="p-6 space-y-6">
